# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  owner: "owner",
  manager: "manager",
} as const;

export interface Hotel {
  id: number;
  name: string;
  totalRooms: number;
  createdAt: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  hotelId?: number | null;
  hotel?: Hotel | null;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  owner: "owner",
  manager: "manager",
} as const;

export interface CreateUserRequest {
  email: string;
  name: string;
  password: string;
  role: CreateUserRequestRole;
  hotelId?: number | null;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  owner: "owner",
  manager: "manager",
} as const;

export interface UpdateUserRequest {
  email?: string;
  name?: string;
  role?: UpdateUserRequestRole;
  hotelId?: number | null;
  password?: string;
}

export interface CreateHotelRequest {
  name: string;
  totalRooms: number;
}

export interface Agency {
  id: number;
  name: string;
  contactEmail?: string | null;
  contactPhone?: string | null;
  hotelId: number;
  createdAt: string;
}

export interface CreateAgencyRequest {
  name: string;
  contactEmail?: string | null;
  contactPhone?: string | null;
  hotelId: number;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  confirmed: "confirmed",
  checked_in: "checked_in",
  checked_out: "checked_out",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  guestName: string;
  guestEmail?: string | null;
  guestPhone?: string | null;
  roomNumber?: string | null;
  roomType?: string | null;
  checkIn: string;
  checkOut: string;
  roomRent: number;
  addOns: number;
  totalCost: number;
  receipt: number;
  balance: number;
  notes?: string | null;
  status: BookingStatus;
  hotelId: number;
  agencyId?: number | null;
  agency?: Agency | null;
  hotel?: Hotel | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateBookingRequestStatus =
  (typeof CreateBookingRequestStatus)[keyof typeof CreateBookingRequestStatus];

export const CreateBookingRequestStatus = {
  confirmed: "confirmed",
  checked_in: "checked_in",
  checked_out: "checked_out",
  cancelled: "cancelled",
} as const;

export interface CreateBookingRequest {
  guestName: string;
  guestEmail?: string | null;
  guestPhone?: string | null;
  roomNumber?: string | null;
  roomType?: string | null;
  checkIn: string;
  checkOut: string;
  roomRent: number;
  addOns: number;
  receipt: number;
  notes?: string | null;
  status: CreateBookingRequestStatus;
  hotelId: number;
  agencyId?: number | null;
}

export type UpdateBookingRequestStatus =
  (typeof UpdateBookingRequestStatus)[keyof typeof UpdateBookingRequestStatus];

export const UpdateBookingRequestStatus = {
  confirmed: "confirmed",
  checked_in: "checked_in",
  checked_out: "checked_out",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingRequest {
  guestName?: string;
  guestEmail?: string | null;
  guestPhone?: string | null;
  roomNumber?: string | null;
  roomType?: string | null;
  checkIn?: string;
  checkOut?: string;
  roomRent?: number;
  addOns?: number;
  receipt?: number;
  notes?: string | null;
  status?: UpdateBookingRequestStatus;
  agencyId?: number | null;
}

export interface UpdatePaymentRequest {
  receipt: number;
}

export interface DashboardStats {
  todayCheckins: number;
  todayCheckouts: number;
  totalBookings: number;
  occupiedRooms: number;
  totalRooms: number;
  occupancyPercentage: number;
  monthlyRevenue: number;
}

export type OccupancyStatsDailyOccupancyItem = {
  date: string;
  occupiedRooms: number;
  totalRooms: number;
  percentage: number;
};

export interface OccupancyStats {
  dailyOccupancy: OccupancyStatsDailyOccupancyItem[];
  averageOccupancy: number;
  totalRooms: number;
}

export type RevenueStatsMonthlyRevenueItem = {
  month: number;
  year: number;
  revenue: number;
  bookings: number;
};

export type RevenueStatsAgencyRevenueItem = {
  agencyId?: number | null;
  agencyName: string;
  revenue: number;
  bookings: number;
};

export interface RevenueStats {
  monthlyRevenue: RevenueStatsMonthlyRevenueItem[];
  agencyRevenue: RevenueStatsAgencyRevenueItem[];
  totalYearlyRevenue: number;
}

export type ListBookingsParams = {
  month?: number;
  year?: number;
  hotelId?: number;
  date?: string;
};

export type GetDashboardStatsParams = {
  hotelId?: number;
};

export type GetTodayCheckinsParams = {
  hotelId?: number;
};

export type GetTodayCheckoutsParams = {
  hotelId?: number;
};

export type GetOccupancyStatsParams = {
  hotelId?: number;
  month?: number;
  year?: number;
};

export type GetRevenueStatsParams = {
  hotelId?: number;
  year?: number;
};