# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hotel CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /agencies:
    get:
      operationId: listAgencies
      tags: [agencies]
      summary: List all agencies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of agencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agency"
    post:
      operationId: createAgency
      tags: [agencies]
      summary: Create an agency
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgencyRequest"
      responses:
        "201":
          description: Agency created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
  /agencies/{id}:
    put:
      operationId: updateAgency
      tags: [agencies]
      summary: Update an agency
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgencyRequest"
      responses:
        "200":
          description: Agency updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
    delete:
      operationId: deleteAgency
      tags: [agencies]
      summary: Delete an agency
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Agency deleted
...
  /analytics/occupancy:
    get:
      operationId: getOccupancyStats
      tags: [analytics]
      summary: Get occupancy analytics
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Occupancy stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccupancyStats"
  /analytics/revenue:
    get:
      operationId: getRevenueStats
      tags: [analytics]
      summary: Get revenue analytics
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, owner, manager]
        hotelId:
          type: integer
          nullable: true
        hotel:
          $ref: "#/components/schemas/Hotel"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
...
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        totalRooms:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - totalRooms
        - createdAt
...
    Agency:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        hotelId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - hotelId
        - createdAt
    CreateAgencyRequest:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        hotelId:
          type: integer
      required:
        - name
        - hotelId
...
    OccupancyStats:
      type: object
      properties:
        dailyOccupancy:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              occupiedRooms:
                type: integer
              totalRooms:
                type: integer
              percentage:
                type: number
            required:
              - date
              - occupiedRooms
              - totalRooms
              - percentage
        averageOccupancy:
          type: number
        totalRooms:
          type: integer
      required:
        - dailyOccupancy
        - averageOccupancy
        - totalRooms
    RevenueStats:
      type: object
      properties:
        monthlyRevenue:
          type: array
          items:
            type: object
            properties:
              month:
                type: integer
              year:
                type: integer
              revenue:
                type: number
              bookings:
                type: integer
            required:
              - month
              - year
              - revenue
              - bookings
        agencyRevenue:
          type: array
          items:
            type: object
            properties:
              agencyId:
                type: integer
                nullable: true
              agencyName:
                type: string
              revenue:
                type: number
              bookings:
                type: integer
            required:
              - agencyName
              - revenue
              - bookings
        totalYearlyRevenue:
          type: number
      required:
        - monthlyRevenue
        - agencyRevenue
        - totalYearlyRevenue