# lib/api-zod/src/generated/types/createAgencyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAgencyRequest {
  name: string;
  contactEmail?: string | null;
  contactPhone?: string | null;
  hotelId: number;
}