# lib/api-zod/src/generated/types/hotel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Hotel {
  id: number;
  name: string;
  totalRooms: number;
  createdAt: Date;
}